static setup of cassandra DAOs for CassandraMailRepositoryMailDAO tests
